функции репутации

+++ bot.py
@@ -87,42 +87,42 @@ def text_handler(message):
 
 # добавление и отнимание репутации
 def reputation(message):
-        connect = sqlite3.connect("data.db")
-        cursor = connect.cursor()
-        cooldown = cursor.execute(f"SELECT cooldown from chat_{str(message.chat.id)[1:]} WHERE id = {message.from_user.id}").fetchone()[0]
+        # проверка правильности написания команды
+        try:
+            to_whom = message.text.split()[1][1:]
+        except IndexError:
+            bot.send_message(message.chat.id, "Ошибка при вводе запроса. Проверьте синтаксис, написав команду /help")
+            return
+
+        if not db.isUserInDBByUsername(message.chat.id, to_whom):
+            bot.send_message(message.chat.id, "Такого пользователя нет в чате или я ещё не знаком с ним. Попросите его написать тут что-то")
+
         # проверка на кулдаун
+        cooldown = db.getCooldown(message.chat.id, message.from_user.id)
         if round(time.time()) - cooldown < 3600:
             cooldown_remain = int(((time.time() - cooldown - 3600) / 60) // -1)
             bot.send_message(message.chat.id, f"Изменять репутацию можно только раз в час! Попробуй снова через {cooldown_remain} минут{'ы' if cooldown_remain in [2, 3, 4] else 'у' if cooldown == 1 else ''}")
-            return "cooldown"
-        to_whom = message.text.split()[1][1:]
+            return
+
+        # репутация бота
         if to_whom == bot.get_me().username:
             if message.text[0] == "-":
                 bot.send_message(message.chat.id, f"Вы решили посягнуть на святое! Я конфисковать у вас {'кошка жена и ' if random.randint(0, 1) == 1 else ''}{random.randint(1, 10)} миска рис!")
             else:
                 bot.send_message(message.chat.id, "Ой спасиба\n   🥺\n👉🏻 👈🏻")
-            return "rep bot"
-        if message.from_user.username != to_whom:
-            cursor.execute(f"SELECT username, reputation from chat_{str(message.chat.id)[1:]}")
-            current = cursor.fetchall()
-            for user in current:
-                if user[0] == to_whom:
-                    rep = user[1] + 1 if message.text[0] == "+" else user[1] - 1
-                    break
-            try:
-                cursor.execute(f"""UPDATE chat_{str(message.chat.id)[1:]} set reputation = ? where username = ?""", (rep, to_whom))
-            except UnboundLocalError:
-                bot.send_message(message.chat.id, "Такого пользователя нет в чате или он не написал /start")
-                return UnboundLocalError
-            cursor.execute(f"""UPDATE chat_{str(message.chat.id)[1:]} set cooldown = {time.time()} where id = {message.from_user.id}""")
-            connect.commit()
+            return
+        # репутация себе
+        elif message.from_user.username == to_whom and message.text[0] in ["+", "-"]:
+            bot.send_message(message.chat.id, f"Нельзя {'повыcить' if message.text[0] == '+' else 'понизить'} репутацию самому себе!")
+        # репутация другому (так и надо)
+        elif message.from_user.username != to_whom and message.text[0] in ["+", "-"]:
+            db.updateReputation(message.chat.id, to_whom, message.text[0])
+            db.setCooldown(message.chat.id, to_whom, time.time())
             bot.delete_message(message.chat.id, message.message_id)
             try:
-                bot.send_message(message.chat.id, f"@{message.from_user.username} {'повышает' if message.text[0] == '+' else 'понижает'} репутацию @{to_whom}.\nПричина: {message.text.split(' ', 2)[2]}.\nТеперь репутация равна {rep}")
+                bot.send_message(message.chat.id, f"@{message.from_user.username} {'повышает' if message.text[0] == '+' else 'понижает'} репутацию @{to_whom}.\nПричина: {message.text.split(' ', 2)[2]}.")
             except IndexError:
-                bot.send_message(message.chat.id, f"@{message.from_user.username} {'повышает' if message.text[0] == '+' else 'понижает'} репутацию @{to_whom}.\nПричина: нет.\nТеперь репутация равна {rep}")
-        else:
-            bot.send_message(message.chat.id, f"Нельзя {'повыcить' if message.text[0] == '+' else 'понизить'} репутацию самому себе!")
+                bot.send_message(message.chat.id, f"@{message.from_user.username} {'повышает' if message.text[0] == '+' else 'понижает'} репутацию @{to_whom}.\nПричина: нет.")
 
 def mooseMeme(message):
     # FIXME: множественное число, слово слитно->пунктуация->лось. "он(лось)"
